webapp/jupyter: wip for lifting up usage info
