+++ src/smc-webapp/jupyter/main.tsx
@@ -39,9 +39,15 @@ import { RawEditor } from "./raw-editor";
 // import { SnippetsDialog } from "smc-webapp/assistant/dialog";
 const { SnippetsDialog } = require("smc-webapp/assistant/dialog");
 import { Kernel as KernelType, Kernels as KernelsType } from "./util";
-import { Scroll } from "./types";
+import { Scroll, Usage, BackendState } from "./types";
+import { ImmutableUsageInfo } from "../../smc-project/usage-info/types";
+import { compute_usage } from "./usage";
 
-const KERNEL_STYLE: React.CSSProperties = {
+import { JupyterActions } from "./browser-actions";
+import { NotebookFrameActions } from "../frame-editors/jupyter-editor/cell-notebook/actions";
+import { JupyterEditorActions } from "../frame-editors/jupyter-editor/actions";
+
+const KERNEL_STYLE: CSS = {
   float: "right",
   paddingLeft: "5px",
   backgroundColor: "#eee",
@@ -50,20 +56,50 @@ const KERNEL_STYLE: React.CSSProperties = {
   borderLeft: "1px solid rgb(231,231,231)",
   borderBottom: "1px solid rgb(231,231,231)",
   whiteSpace: "nowrap",
-};
+} as const;
 
-import { JupyterActions } from "./browser-actions";
-import { NotebookFrameActions } from "../frame-editors/jupyter-editor/cell-notebook/actions";
-import { JupyterEditorActions } from "../frame-editors/jupyter-editor/actions";
-
-const ERROR_STYLE = {
+const ERROR_STYLE: CSS = {
   margin: "1ex",
   whiteSpace: "pre" as "pre",
   fontSize: "12px",
   fontFamily: "monospace" as "monospace",
   maxHeight: "30vh",
   overflow: "auto",
-} as CSS;
+} as const;
+
+// derive sorted list of timings from all cells
+function calc_cell_timings(cells?: immutable.Map<string, any>): number[] {
+  if (cells == null) return [];
+  return cells
+    .toList()
+    .map((v) => {
+      const start = v.get("start");
+      const end = v.get("end");
+      if (start != null && end != null) {
+        return (end - start) / 1000;
+      } else {
+        return null;
+      }
+    })
+    .filter((v) => v != null)
+    .sort()
+    .toJS();
+}
+
+// for the sorted list of cell timing, get the median or quantile.
+// a quick approximation is good enough for us!
+// we basically want to ignore long running cells, treat them as outliers.
+// Using the 75% quantile is quick and easy, avoids working with inter quantile differences
+// and proper outlier detection – like for boxplots, etc.
+// we also cap the lower end with a reasonable minimum.
+// Maybe another choice of quantile works better, something for later …
+function calc_quantile(data: number[], min_val = 3, q = 0.75): number {
+  if (data.length == 0) return min_val;
+  const idx_last = data.length - 1;
+  const idx_q = Math.floor(q * idx_last);
+  const idx = Math.min(idx_last, idx_q);
+  return Math.max(min_val, data[idx]);
+}
 
 interface Props {
   error?: string;
@@ -237,6 +273,69 @@ export const JupyterEditor: React.FC<Props> = React.memo((props: Props) => {
   //   "kernel_streams",
   // ]);
   const kernel_error: undefined | string = useRedux([name, "kernel_error"]);
+  const kernel_usage: undefined | ImmutableUsageInfo = useRedux([
+    name,
+    "kernel_usage",
+  ]);
+  const backend_state: undefined | BackendState = useRedux([
+    name,
+    "backend_state",
+  ]);
+  const kernel_state: undefined | string = useRedux([name, "kernel_state"]);
+
+  // cell timing statistic
+  const cell_timings = React.useMemo(() => calc_cell_timings(cells), [cells]);
+  const expected_cell_runtime = React.useMemo(
+    () => calc_quantile(cell_timings),
+    [cell_timings]
+  );
+  console.log("expected_cell_runtime", expected_cell_runtime);
+
+  // state of UI, derived from usage, timing stats, etc.
+  const [cpu_start, set_cpu_start] = React.useState<number | undefined>();
+  const [cpu_runtime, set_cpu_runtime] = React.useState<number>(0);
+  const timer1 = React.useRef<ReturnType<typeof setInterval> | undefined>();
+
+  // reset cpu_start time when state changes
+  React.useEffect(() => {
+    if (kernel_state == "busy") {
+      set_cpu_start(Date.now());
+    } else if (cpu_start != null) {
+      set_cpu_start(undefined);
+    }
+  }, [kernel_state]);
+
+  // count seconds when kernel is busy & reset counter
+  React.useEffect(() => {
+    if (cpu_start != null) {
+      timer1.current = setInterval(() => {
+        if (kernel_state == "busy") {
+          set_cpu_runtime((Date.now() - cpu_start) / 1000);
+        } else {
+          set_cpu_runtime(0);
+        }
+      }, 100);
+    } else if (timer1.current != null) {
+      set_cpu_runtime(0);
+      clearInterval(timer1.current);
+    }
+    return () => {
+      if (timer1.current != null) clearInterval(timer1.current);
+    };
+  }, [cpu_start, kernel_state]);
+
+  // based on the info we know, we derive the "usage" object
+  // the "status.tsx" Kernel component and other UI details will visualize it
+  const usage: Usage = React.useMemo(
+    () =>
+      compute_usage({
+        kernel_usage,
+        backend_state,
+        cpu_runtime,
+        expected_cell_runtime,
+      }),
+    [kernel_usage, backend_state, cpu_runtime, expected_cell_runtime]
+  );
 
   function render_kernel_error() {
     if (!kernel_error) return;
@@ -281,7 +380,9 @@ export const JupyterEditor: React.FC<Props> = React.memo((props: Props) => {
           is_fullscreen={is_fullscreen}
           name={name}
           actions={actions}
-          cells={cells}
+          usage={usage}
+          cpu_runtime={cpu_runtime}
+          expected_cell_runtime={expected_cell_runtime}
         />
         <Mode name={name} />
       </span>
@@ -330,6 +431,7 @@ export const JupyterEditor: React.FC<Props> = React.memo((props: Props) => {
           cur_id={cur_id}
           sel_ids={sel_ids}
           cell_toolbar={cell_toolbar}
+          usage={usage}
         />
       );
     }
