+++ src/smc-webapp/jupyter/types.ts
@@ -31,3 +31,24 @@ export type CellToolbarName =
 export type Cell = immutable.Map<string, any>;
 
 export type Cells = immutable.Map<string, Cell>;
+
+export interface Usage {
+  mem: number; // MiB
+  mem_limit: number;
+  mem_alert: AlertLevel;
+  mem_pct: number; // %
+  cpu: number; // 1 core = 100%
+  cpu_limit: number;
+  cpu_alert: AlertLevel;
+  cpu_pct: number; // 100% full container quota
+  time_alert: AlertLevel;
+}
+
+export type AlertLevel = "low" | "mid" | "high" | "none";
+
+export type BackendState =
+  | "init"
+  | "ready"
+  | "spawning"
+  | "starting"
+  | "running";
